The default argument of LS::FileSystem has changed to . (cwd)
It caused issues on devices like Android. It makes perfect sense to add . instead of /. If LinuxStat is on your thumbdrive for example, it will show the info about the thumbdrive and not about the / of the system. This new behaviour won't cause issues, since the app installed is most likely to be under /home/user, which is under /.