Updated battery.rb
LS::Battery has new methods like devices_stat() which shows all device's stat that has a battery. It can be a keyboard, mouse or the battery of a system itself. It could also be dual batteries of a system. It's slow compared to the other methods because it opens multiple files multiple times. But in situations, this can be very handy little method. charge_full_wh and charge_full_design_wh is also added, which is shown by various utilies but was missing in LinuxStat. Other methods has some minor performance improvements when called in a loop.