Performance improvements for process.rb
The previous Process module was super slow just because it wasn't implemented nicely. It used everything that would slow down the application. They are now removed and it's now a bit more optimized, at least it takes ~0.09 seconds to detect 4000 zombie processes out of 4200 processes. Previously it used to take ~0.2 seconds